add display result method for dice result scene

+++ src/main/java/controller/tkfisch/Controller.java
@@ -68,7 +68,7 @@ public class Controller {
             System.out.println("Scene not found: " + name);
         }
     }
-    public void playAnimation(String imgURL, int frames, int duration, ImageView animatingObj){
+    public void playAnimation(String imgURL, int frames, double duration, ImageView animatingObj){
         //play a single cycle of animation on a given obj
         for (int i =0; i<= frames; i++){
             Image img;
